+++ HappyHouseServer/src/main/java/com/ssafy/happyhouse/model/mapper/HouseMapper.java
@@ -1,5 +1,6 @@
 package com.ssafy.happyhouse.model.mapper;
 
+import java.math.BigInteger;
 import java.sql.SQLException;
 import java.util.List;
 
@@ -9,5 +10,5 @@ import com.ssafy.happyhouse.model.HouseDealDto;
 
 @Mapper
 public interface HouseMapper {
-	List<HouseDealDto> getHouseDealInAptNo(int aptNo) throws SQLException;
+	List<HouseDealDto> getHouseDealInAptNo(BigInteger aptNo) throws SQLException;
 }
